test: fold the route-body type assertions into the typecheck test
The separate assertion module was a leftover of the first design, where a
dedicated tsconfig compiled only it. Since the compile grew to cover every
test source, the runner's own file is part of the project it compiles, so
the assertions live beside the spawn and the file is named for what it
does. Two files and a directory become one.
